Removed unneeded PoliException; improved Exception defs which are Serializable; Back down to SLF5J 1.5 (for Java6 runtime issues);  Heed the redundant casting calls long-to-int, int-to-int, etc

+++ Core/src/main/java/org/opensextant/extraction/TextEntity.java
@@ -108,7 +108,7 @@ public class TextEntity {
             // Match not initialized
             return 0;
         }
-        return (int) (end - start);
+        return (end - start);
     }
 
     /** Convenience methods for carrying the context through the output processing */
